third_party/blink/web_tests/svg/dynamic-updates/SVGFEDistantLightElement-dom-azimuth-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'azimuth' attribute of the SVGFEDistantLightElement object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS distantLight.getAttribute('azimuth') is "10"
PASS distantLight.getAttribute('azimuth') is "100"
PASS successfullyParsed is true

TEST COMPLETE
